server/requirements.txt
flask
flask-login
flask-sqlalchemy
flask-cors
pytest
psycopg2
